# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/catalog/products/{slug}": {
      "get": {
        "operationId": "CatalogController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/admin/products": {
      "get": {
        "operationId": "AdminCatalogController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      },
      "post": {
        "operationId": "AdminCatalogController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sku": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._\\-/]{0,63}$"
                  },
                  "slug": {
                    "type": "string",
                    "maxLength": 120,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                  },
                  "nameFa": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "nameEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 200
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "aliases": {
                    "default": [],
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    }
                  },
                  "descriptionFa": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 5000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "descriptionEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 5000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "manufacturerBrandId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "vehicleBrandIds": {
                    "default": [],
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "partType": {
                    "type": "string",
                    "enum": [
                      "GENUINE",
                      "OEM",
                      "AFTERMARKET"
                    ]
                  },
                  "condition": {
                    "type": "string",
                    "enum": [
                      "NEW",
                      "USED",
                      "REFURBISHED"
                    ]
                  },
                  "origin": {
                    "type": "string",
                    "enum": [
                      "DOMESTIC",
                      "IMPORTED",
                      "UNKNOWN"
                    ]
                  },
                  "countryOfManufacture": {
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "warrantyFa": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 500
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "warrantyEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 500
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "unit": {
                    "default": "عدد",
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 30
                  },
                  "packQuantity": {
                    "default": 1,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10000
                  },
                  "oemCodes": {
                    "default": [],
                    "maxItems": 30,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 64
                    }
                  },
                  "specs": {
                    "default": [],
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "maxLength": 64
                        },
                        "labelFa": {
                          "type": "string",
                          "maxLength": 100
                        },
                        "labelEn": {
                          "type": "string",
                          "maxLength": 100
                        },
                        "value": {
                          "type": "string",
                          "maxLength": 300
                        }
                      },
                      "required": [
                        "key",
                        "labelFa",
                        "value"
                      ]
                    }
                  },
                  "basePrice": {
                    "type": "object",
                    "properties": {
                      "currency": {
                        "type": "string",
                        "enum": [
                          "IRR",
                          "AED"
                        ]
                      },
                      "amountMinor": {
                        "type": "string",
                        "pattern": "^\\d{1,20}$"
                      }
                    },
                    "required": [
                      "currency",
                      "amountMinor"
                    ]
                  },
                  "manualPriceIrr": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{1,20}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "manualPriceAed": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{1,20}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "lowStockThreshold": {
                    "default": 0,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000
                  },
                  "isSellable": {
                    "default": true,
                    "type": "boolean"
                  },
                  "version": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "sku",
                  "nameFa",
                  "categoryId",
                  "partType",
                  "condition",
                  "origin",
                  "basePrice"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      }
    },
    "/api/v1/admin/products/{id}": {
      "get": {
        "operationId": "AdminCatalogController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      },
      "put": {
        "operationId": "AdminCatalogController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sku": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._\\-/]{0,63}$"
                  },
                  "slug": {
                    "type": "string",
                    "maxLength": 120,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                  },
                  "nameFa": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "nameEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 200
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "aliases": {
                    "default": [],
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    }
                  },
                  "descriptionFa": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 5000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "descriptionEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 5000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "manufacturerBrandId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "vehicleBrandIds": {
                    "default": [],
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "partType": {
                    "type": "string",
                    "enum": [
                      "GENUINE",
                      "OEM",
                      "AFTERMARKET"
                    ]
                  },
                  "condition": {
                    "type": "string",
                    "enum": [
                      "NEW",
                      "USED",
                      "REFURBISHED"
                    ]
                  },
                  "origin": {
                    "type": "string",
                    "enum": [
                      "DOMESTIC",
                      "IMPORTED",
                      "UNKNOWN"
                    ]
                  },
                  "countryOfManufacture": {
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "warrantyFa": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 500
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "warrantyEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 500
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "unit": {
                    "default": "عدد",
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 30
                  },
                  "packQuantity": {
                    "default": 1,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10000
                  },
                  "oemCodes": {
                    "default": [],
                    "maxItems": 30,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 64
                    }
                  },
                  "specs": {
                    "default": [],
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "maxLength": 64
                        },
                        "labelFa": {
                          "type": "string",
                          "maxLength": 100
                        },
                        "labelEn": {
                          "type": "string",
                          "maxLength": 100
                        },
                        "value": {
                          "type": "string",
                          "maxLength": 300
                        }
                      },
                      "required": [
                        "key",
                        "labelFa",
                        "value"
                      ]
                    }
                  },
                  "basePrice": {
                    "type": "object",
                    "properties": {
                      "currency": {
                        "type": "string",
                        "enum": [
                          "IRR",
                          "AED"
                        ]
                      },
                      "amountMinor": {
                        "type": "string",
                        "pattern": "^\\d{1,20}$"
                      }
                    },
                    "required": [
                      "currency",
                      "amountMinor"
                    ]
                  },
                  "manualPriceIrr": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{1,20}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "manualPriceAed": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{1,20}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "lowStockThreshold": {
                    "default": 0,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000
                  },
                  "isSellable": {
                    "default": true,
                    "type": "boolean"
                  },
                  "version": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "sku",
                  "nameFa",
                  "categoryId",
                  "partType",
                  "condition",
                  "origin",
                  "basePrice"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      }
    },
    "/api/v1/admin/products/{id}/publish": {
      "post": {
        "operationId": "AdminCatalogController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      }
    },
    "/api/v1/admin/products/{id}/archive": {
      "post": {
        "operationId": "AdminCatalogController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      }
    },
    "/api/v1/admin/products/{id}/media": {
      "post": {
        "operationId": "AdminCatalogController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      }
    },
    "/api/v1/admin/products/{id}/media/{mediaId}": {
      "patch": {
        "operationId": "AdminCatalogController_updateMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      },
      "delete": {
        "operationId": "AdminCatalogController_removeMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      }
    },
...
    "/api/v1/admin/products/{id}/price-rules/{ruleId}": {
      "delete": {
        "operationId": "AdminCatalogController_removeRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ruleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "HEDAX API",
    "description": "Auto parts store and custom sourcing platform — REST API v1. Money is always { currency, amountMinor: string }.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hedax_sid"
      }
    },
...
  }
}